[py] fix: code review changes
- moved virtual_authenticator to common
- moved virtual auth tests to virtual_authenticator dir
- mark xfail for safari and firefox
