+++ py/selenium/webdriver/remote/command.py
@@ -87,109 +87,6 @@ class Command(object):
     MINIMIZE_WINDOW = "minimizeWindow"
     PRINT_PAGE = 'printPage'
 
-    GET_SESSION_LOGS = "getSessionLogs"
-    GET_ALL_SESSIONS = "getAllSessions"
-    GET_CAPABILITIES = "getCapabilities"
-    STATUS = "status"
-    GET_ELEMENT_SHADOW_ROOT = "getElementShadowRoot"
-    FIND_ELEMENT_FROM_SHADOW_ROOT = "findElementFromShadowRoot"
-    FIND_ELEMENTS_FROM_SHADOW_ROOT = "findElementsFromShadowRoot"
-    SEND_KEYS_TO_ACTIVE_ELEMENT = "sendKeysToActiveElement"
-    GET_CURRENT_WINDOW_HANDLE = "getCurrentWindowHandle"
-    GET_WINDOW_HANDLES = "getWindowHandles"
-
-    GET_CURRENT_CONTEXT_HANDLE = "getCurrentContextHandle"
-    GET_CONTEXT_HANDLES = "getContextHandles"
-
-    SWITCH_TO_NEW_WINDOW = "newWindow"
-    SWITCH_TO_CONTEXT = "switchToContext"
-    GET_ACTIVE_ELEMENT = "getActiveElement"
-
-    EXECUTE_SCRIPT = "executeScript"
-    IS_ELEMENT_DISPLAYED = "isElementDisplayed"
-    GET_ELEMENT_LOCATION = "getElementLocation"
-    GET_ELEMENT_LOCATION_ONCE_SCROLLED_INTO_VIEW = "getElementLocationOnceScrolledIntoView"
-    GET_ELEMENT_SIZE = "getElementSize"
-    GET_ELEMENT_DOM_PROPERTY = "getElementDomProperty"
-    GET_ELEMENT_DOM_ATTRIBUTE = "getElementDomAttribute"
-    GET_ELEMENT_ACCESSIBLE_NAME = "getElementAccessibleName"
-    ELEMENT_EQUALS = "elementEquals"
-
-    ACCEPT_ALERT = "acceptAlert"
-    DISMISS_ALERT = "dismissAlert"
-    GET_ALERT_TEXT = "getAlertText"
-    SET_ALERT_VALUE = "setAlertValue"
-    SET_ALERT_CREDENTIALS = "setAlertCredentials"
-
-    IMPLICITLY_WAIT = "implicitlyWait"
-    SET_SCRIPT_TIMEOUT = "setScriptTimeout"
-
-    GET_LOCATION = "getLocation"
-    SET_LOCATION = "setLocation"
-    GET_APP_CACHE = "getAppCache"
-    GET_APP_CACHE_STATUS = "getStatus"
-    CLEAR_APP_CACHE = "clearAppCache"
-    IS_BROWSER_ONLINE = "isBrowserOnline"
-    SET_BROWSER_ONLINE = "setBrowserOnline"
-
-    GET_LOCAL_STORAGE_ITEM = "getLocalStorageItem"
-    GET_LOCAL_STORAGE_KEYS = "getLocalStorageKeys"
-    SET_LOCAL_STORAGE_ITEM = "setLocalStorageItem"
-    REMOVE_LOCAL_STORAGE_ITEM = "removeLocalStorageItem"
-    CLEAR_LOCAL_STORAGE = "clearLocalStorage"
-    GET_LOCAL_STORAGE_SIZE = "getLocalStorageSize"
-
-    GET_SESSION_STORAGE_ITEM = "getSessionStorageItem"
-    GET_SESSION_STORAGE_KEYS = "getSessionStorageKey"
-    SET_SESSION_STORAGE_ITEM = "setSessionStorageItem"
-    REMOVE_SESSION_STORAGE_ITEM = "removeSessionStorageItem"
-    CLEAR_SESSION_STORAGE = "clearSessionStorage"
-    GET_SESSION_STORAGE_SIZE = "getSessionStorageSize"
-
-    SET_SCREEN_ORIENTATION = "setScreenOrientation"
-    GET_SCREEN_ORIENTATION = "getScreenOrientation"
-    SET_SCREEN_ROTATION = "setScreenRotation"
-    GET_SCREEN_ROTATION = "getScreenRotation"
-
-    # W3C Actions APIs
-    ACTIONS = "actions"
-    CLEAR_ACTIONS_STATE = "clearActionsState"
-
-    # These belong to the Advanced user interactions - an element is optional for these commands
-    CLICK = "mouseClick"
-    DOUBLE_CLICK = "mouseDoubleClick"
-    MOUSE_DOWN = "mouseButtonDown"
-    MOUSE_UP = "mouseButtonUp"
-    MOVE_TO = "mouseMoveTo"
-
-    # Those allow interactions with the Input Methods installed on the system.
-    IME_GET_AVAILABLE_ENGINES = "imeGetAvailableEngines"
-    IME_GET_ACTIVE_ENGINE = "imeGetActiveEngine"
-    IME_IS_ACTIVATED = "imeIsActivated"
-    IME_DEACTIVATE = "imeDeactivate"
-    IME_ACTIVATE_ENGINE = "imeActivateEngine"
-
-    # Advanced touch API
-    TOUCH_SINGLE_TAP = "touchSingleTap"
-    TOUCH_DOWN = "touchDown"
-    TOUCH_UP = "touchUp"
-    TOUCH_MOVE = "touchMove"
-    TOUCH_SCROLL = "touchScroll"
-    TOUCH_DOUBLE_TAP = "touchDoubleTap"
-    TOUCH_LONG_PRESS = "touchLongPress"
-    TOUCH_FLICK = "touchFlick"
-
-    # Window API
-    SET_CURRENT_WINDOW_POSITION = "setWindowPosition"
-    GET_CURRENT_WINDOW_POSITION = "getWindowPosition"
-
-    # W3C compatible Window API
-    SET_CURRENT_WINDOW_SIZE = "setCurrentWindowSize"
-    GET_CURRENT_WINDOW_SIZE = "getCurrentWindowSize"
-    MAXIMIZE_CURRENT_WINDOW = "maximizeCurrentWindow"
-    MINIMIZE_CURRENT_WINDOW = "minimizeCurrentWindow"
-    FULLSCREEN_CURRENT_WINDOW = "fullscreenCurrentWindow"
-
     # Alerts
     W3C_DISMISS_ALERT = "w3cDismissAlert"
     W3C_ACCEPT_ALERT = "w3cAcceptAlert"
