+++ py/test/unit/selenium/webdriver/virtual_authenticator/virtual_authenticator_options_tests.py
@@ -18,7 +18,7 @@
 import pytest
 
 
-from selenium.webdriver.chrome.virtual_authenticator import (
+from selenium.webdriver.common.virtual_authenticator import (
     VirtualAuthenticatorOptions,
 )
